Separating RTT logic from Brainclouds2s into BrainClouds2sRTTComms
Adjusted the test to also include DisableRTT to show its all working.
