+++ unityS2STest/s2sTest/Assets/s2s_src/Brainclouds2sRttComms.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 44e102c91ade3754099ed03ca47e733c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
